Inline the EmbargoConcern
So we can use the attribute class method to define fields
