+++ app/models/solr_document.rb
@@ -9,8 +9,17 @@ class SolrDocument
   include TitleConcern
   include DocumentDateConcern
   include PreservationSizeConcern
-  include EmbargoConcern
   include ObjectTypeConcern
+
+  FIELD_EMBARGO_STATUS       = :embargo_status_ssim
+  FIELD_EMBARGO_RELEASE_DATE = :embargo_release_dtsim
+
+  attribute :embargo_status, Blacklight::Types::String, FIELD_EMBARGO_STATUS
+  attribute :embargo_release_date, Blacklight::Types::String, FIELD_EMBARGO_RELEASE_DATE
+
+  def embargoed?
+    embargo_status == 'embargoed'
+  end
 
   # self.unique_key = 'id'
